pr-fix: revert hiding test functionality
